refactor(mirv_pov): isolate death panel lifecycle
